feat: atomic design の templates 層を導入し全ページから抽出
- app/components/templates/ に 9 つのテンプレートコンポーネントを作成
  - HomeTemplate, LoginTemplate, UserRegisterTemplate
  - ListeningLogsTemplate, ListeningLogNewTemplate
  - ListeningLogDetailTemplate, ListeningLogEditTemplate
  - PiecesTemplate, PieceNewTemplate, PieceEditTemplate
- 各 page はデータ取得・ロジックのみを担い、template に props/emit で委譲
- nuxt.config.ts に pathPrefix: false を追加し subdirectory なしで参照可能に
