+++ app/components/templates/ListeningLogDetailTemplate.vue
@@ -1,0 +1,116 @@
+<script setup lang="ts">
+import { formatDatetime } from "~/utils/date";
+import type { ListeningLog } from "~/types";
+
+defineProps<{
+  log: ListeningLog;
+}>();
+
+const { ratingStars } = useRatingDisplay();
+</script>
+
+<template>
+  <div>
+    <div class="page-header">
+      <NuxtLink to="/listening-logs" class="back-link">← 鑑賞記録一覧</NuxtLink>
+      <NuxtLink :to="`/listening-logs/${log.id}/edit`" class="btn-secondary">編集</NuxtLink>
+    </div>
+
+    <article class="log-detail">
+      <header>
+        <h1>
+          <span v-if="log.isFavorite" class="favorite">♥</span>
+          {{ log.piece }}
+        </h1>
+      </header>
+
+      <dl class="detail-list">
+        <dt>作曲家</dt>
+        <dd>{{ log.composer }}</dd>
+
+        <dt>鑑賞日時</dt>
+        <dd>{{ formatDatetime(log.listenedAt) }}</dd>
+
+        <dt>評価</dt>
+        <dd class="rating">{{ ratingStars(log.rating) }}</dd>
+      </dl>
+
+      <section v-if="log.memo" class="memo">
+        <h2>感想・メモ</h2>
+        <p>{{ log.memo }}</p>
+      </section>
+    </article>
+  </div>
+</template>
+
+<style scoped>
+.page-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 1.5rem;
+}
+
+.back-link {
+  color: #666;
+  text-decoration: none;
+  font-size: 0.9rem;
+}
+
+.back-link:hover {
+  color: #333;
+}
+
+.log-detail {
+  background: #f2e6c9;
+  border: 1px solid #b8995e;
+  border-radius: 12px;
+  padding: 2rem;
+  max-width: 720px;
+}
+
+.log-detail header {
+  margin-bottom: 1.5rem;
+  padding-bottom: 1rem;
+  border-bottom: 1px solid #b8995e;
+}
+
+.log-detail h1 {
+  font-size: 1.8rem;
+  color: #1a1a2e;
+}
+
+.favorite {
+  color: #e05a5a;
+}
+
+.detail-list {
+  display: grid;
+  grid-template-columns: 8rem 1fr;
+  gap: 0.6rem 1rem;
+  margin-bottom: 1.5rem;
+}
+
+dt {
+  font-weight: bold;
+  color: #888;
+  font-size: 0.9rem;
+}
+
+.rating {
+  color: #c9a227;
+  font-size: 1.1rem;
+  letter-spacing: 2px;
+}
+
+.memo h2 {
+  font-size: 1rem;
+  color: #888;
+  margin-bottom: 0.5rem;
+}
+
+.memo p {
+  line-height: 1.7;
+  color: #444;
+}
+</style>
