+++ app/pages/listening-logs/[id]/edit.vue
@@ -20,22 +20,5 @@ async function handleSubmit(values: UpdateListeningLogInput) {
 </script>
 
 <template>
-  <div v-if="log">
-    <h1 class="page-title">鑑賞記録を編集</h1>
-    <p v-if="error" class="error-message">{{ error }}</p>
-    <ListeningLogForm :initial-values="log" submit-label="更新する" @submit="handleSubmit" />
-  </div>
+  <ListeningLogEditTemplate v-if="log" :log="log" :error="error" @submit="handleSubmit" />
 </template>
-
-<style scoped>
-.page-title {
-  font-size: 1.6rem;
-  color: #1a1a2e;
-  margin-bottom: 1.5rem;
-}
-
-.error-message {
-  color: #e05a5a;
-  margin-bottom: 1rem;
-}
-</style>
