+++ app/pages/pieces/new.vue
@@ -2,7 +2,6 @@
 import type { CreatePieceInput } from "~/types";
 
 const { createPiece } = usePieces();
-
 const errorMessage = ref("");
 
 async function handleSubmit(values: CreatePieceInput) {
@@ -17,29 +16,5 @@ async function handleSubmit(values: CreatePieceInput) {
 </script>
 
 <template>
-  <div>
-    <h1 class="page-title">楽曲を追加</h1>
-
-    <div v-if="errorMessage" class="error-message">{{ errorMessage }}</div>
-
-    <PieceForm submit-label="登録する" @submit="handleSubmit" />
-  </div>
+  <PieceNewTemplate :error-message="errorMessage" @submit="handleSubmit" />
 </template>
-
-<style scoped>
-.page-title {
-  font-size: 1.6rem;
-  color: #1a1a2e;
-  margin-bottom: 1.5rem;
-}
-
-.error-message {
-  background: #fff0f0;
-  border: 1px solid #f5c6c6;
-  color: #c0392b;
-  padding: 0.8rem 1rem;
-  border-radius: 6px;
-  margin-bottom: 1rem;
-  font-size: 0.9rem;
-}
-</style>
